% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_plots_to_tmp.R
\name{igraph_plots_to_tmp}
\alias{igraph_plots_to_tmp}
\title{importFrom("grDevices", "dev.off", "jpeg")
importFrom("graphics", "plot")}
\usage{
igraph_plots_to_tmp(igraphs, open = FALSE)
}
\description{
importFrom("grDevices", "dev.off", "jpeg")
importFrom("graphics", "plot")
}
